+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/data/migration/migrations/CategoryLayoutMigration.kt
@@ -1,0 +1,30 @@
+package ch.rmy.android.http_shortcuts.data.migration.migrations
+
+import com.google.gson.JsonObject
+import io.realm.DynamicRealm
+
+object CategoryLayoutMigration : BaseMigration {
+
+    override fun migrateImport(base: JsonObject) {
+        base.getAsJsonArray("categories")
+            ?.map { it.asJsonObject }
+            ?.forEach { category ->
+                when (category.get("layoutType")?.takeUnless { it.isJsonNull }?.asString) {
+                    "grid" -> category.addProperty("layoutType", "dense_grid")
+                }
+            }
+    }
+
+    override fun migrateRealm(realm: DynamicRealm) {
+        realm.where("Category")
+            .findAll()
+            .forEach { category ->
+                when (category.getString("layoutType")) {
+                    "grid" -> category.setString("layoutType", "dense_grid")
+                }
+            }
+    }
+
+    override val version: Int
+        get() = 51
+}